+++ clang/test/Preprocessor/backslash_without_newline.h
@@ -1,0 +1,4 @@
+#define BACKSLASH_WITH_NEWLINE \
+
+// This file intentionally does not end with a newline.
+#define BACKSLASH_WITHOUT_NEWLINE \